+++ Solutions/CrowdStrike Falcon Endpoint Protection/Analytic Rules/CriticalOrHighSeverityDetectionsByUser.yaml
@@ -3,6 +3,7 @@ name: Critical or High Severity Detections by User
 description: |
   'Creates an incident when a large number of Critical/High severity CrowdStrike Falcon sensor detections is triggered by a single user'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: CrowdStrikeFalconEndpointProtection
     dataTypes:
