+++ Solutions/CrowdStrike Falcon Endpoint Protection/Analytic Rules/CriticalSeverityDetection.yaml
@@ -3,6 +3,7 @@ name: Critical Severity Detection
 description: |
   'Creates an incident when a CrowdStrike Falcon sensor detection is triggered with a Critical Severity'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: CrowdStrikeFalconEndpointProtection
     dataTypes:
