+++ Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/Connector_Syslog_CrowdStrikeFalconEndpointProtection.json
