+++ Solutions/CrowdStrike Falcon Endpoint Protection/Data/Solution_CrowdStrike.json
@@ -1,0 +1,25 @@
+{
+  "Name": "CrowdStrike Falcon Endpoint Protection",
+  "Author": "Microsoft - support@microsoft.com",
+  "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Solutions/CrowdStrike%20Falcon%20Endpoint%20Protection/DataConnectors/Logo/crowdstrike.svg\" width=\"75px\" height=\"75px\">",
+  "Description": "The [CrowdStrike Falcon Endpoint Protection](https://www.crowdstrike.com/products/) connector allows you to easily connect your CrowdStrike Falcon Event Stream with Microsoft Sentinel, to create custom dashboards, alerts, and improve investigation. This gives you more insight into your organization's endpoints and improves your security operation capabilities.\n\n**Underlying Microsoft Technologies used:**\n\nThis solution takes a dependency on the following technologies, and some of these dependencies either may be in [Preview](https://azure.microsoft.com/support/legal/preview-supplemental-terms/) state or might result in additional ingestion or operational costs:\n\na. [Agent-based log collection (Syslog)](https://docs.microsoft.com/azure/sentinel/connect-syslog)",
+  "Data Connectors": [
+    "Data Connectors/CrowdstrikeReplicator/CrowdstrikeReplicator_API_FunctionApp.json"
+  ],
+  "Parsers": [
+    "Parsers/CrowdstrikeFalconEventStream.txt",
+	"Parsers/CrowdstrikeReplicator"
+  ],
+   "Workbooks": [
+    "Workbooks/CrowdStrikeFalconEndpointProtection.json"
+  ],
+  "Analytic Rules": [
+    "Analytic Rules/CriticalOrHighSeverityDetectionsByUser.yaml",
+	"Analytic Rules/CriticalSeverityDetection.yaml"
+  ],
+  "BasePath": "C:\\GitHub\\Azure-Sentinel\\Solutions\\CrowdStrike Falcon Endpoint Protection",
+  "Version": "2.0.0",
+  "Metadata": "SolutionMetadata.json",
+  "TemplateSpec": true,
+  "Is1Pconnector": true
+}
