+++ Solutions/CrowdStrike Falcon Endpoint Protection/SolutionMetadata.json
@@ -1,0 +1,20 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-crowdstrikefalconep",
+	"firstPublishDate": "2022-06-01",
+	"providers": ["Crowdstrike"],
+	"categories": {
+		"domains" : ["Security - Threat Protection"],
+		"verticals": []
+	},
+	"support": {
+	  "name": "Microsoft Corporation",
+	  "email": "support@microsoft.com",
+	  "tier": "Microsoft",
+	  "link": "https://support.microsoft.com"
+	}
+}
+
+
+
+
